fix node tests
