--- packages/@glimmer/core/src/host-meta.ts
@@ -1,11 +1,0 @@
-const COMPONENT_INSTANCE_HOST_META_MAP = new WeakMap<{}, unknown>();
-
-export const HOST_META_KEY = `__HOST_META_${Math.floor(Math.random() * Date.now())}__`;
-
-export function getHostMeta(component: object): unknown {
-  return COMPONENT_INSTANCE_HOST_META_MAP.get(component);
-}
-
-export function setHostMeta(component: object, hostMeta: unknown): void {
-  COMPONENT_INSTANCE_HOST_META_MAP.set(component, hostMeta);
-}
