+++ packages/@glimmer/core/test/interactive/index.ts
@@ -1,2 +1,3 @@
+import './render-test';
 import './helper-test';
 import './modifier-test';
