+++ packages/@glimmer/helper/src/helper.ts
@@ -1,17 +1,8 @@
-import { Helper as GlimmerHelper, Dict } from '@glimmer/interfaces';
+import { Helper as GlimmerHelper } from '@glimmer/interfaces';
 import HelperReference, { UserHelper } from './reference';
-import { Reference } from '@glimmer/reference';
-import { HOST_META_KEY } from '@glimmer/core';
 
 export function helper(helperFn: UserHelper): GlimmerHelper {
-  return (args, vm): HelperReference => {
-    const dynamicScope = vm.dynamicScope();
-    let services;
-
-    if (dynamicScope) {
-      services = dynamicScope.get(HOST_META_KEY) as Reference<Dict<unknown>>;
-    }
-
-    return new HelperReference(helperFn, args, services);
+  return (args, _vm): HelperReference => {
+    return new HelperReference(helperFn, args);
   };
 }
