+++ packages/@glimmer/helper/src/reference.ts
@@ -1,38 +1,26 @@
-import { Tag, combine } from '@glimmer/validator';
-import { CachedReference, Reference, VersionedPathReference } from '@glimmer/reference';
+import { Tag } from '@glimmer/validator';
+import { CachedReference, VersionedPathReference } from '@glimmer/reference';
 import { Dict, VMArguments, CapturedArguments } from '@glimmer/interfaces';
 import { NestedPropertyReference } from '@glimmer/core';
 
-export type HelperOptions = {
-  services?: Dict<unknown>;
-};
-
 export type UserHelper = (
   args: ReadonlyArray<unknown>,
-  named: Dict<unknown>,
-  options: HelperOptions
+  named: Dict<unknown>
 ) => unknown;
 
 export default class HelperReference extends CachedReference<unknown>
   implements VersionedPathReference {
   public tag: Tag;
   private args: CapturedArguments;
-  private servicesRef?: Reference<Dict<unknown>>;
 
   constructor(
     private helper: UserHelper,
     args: VMArguments,
-    servicesRef?: Reference<Dict<unknown>>
   ) {
     super();
-    if (servicesRef) {
-      this.tag = combine([args.tag, servicesRef.tag]);
-    } else {
-      this.tag = args.tag;
-    }
+    this.tag = args.tag;
 
     this.args = args.capture();
-    this.servicesRef = servicesRef;
   }
 
   get(key: string): VersionedPathReference {
@@ -41,12 +29,7 @@ export default class HelperReference extends CachedReference<unknown>
 
   compute(): unknown {
     const { helper, args } = this;
-    const options: HelperOptions = {};
 
-    if (this.servicesRef) {
-      options.services = this.servicesRef.value();
-    }
-
-    return helper(args.positional.value(), args.named.value(), options);
+    return helper(args.positional.value(), args.named.value());
   }
 }
