+++ packages/example-apps/basic/index.ts
@@ -1,6 +1,5 @@
 import { renderComponent } from '@glimmer/core';
 import MyComponent from './src/MyComponent';
-import LocaleService from './src/services/LocaleService';
 
 document.addEventListener(
   'DOMContentLoaded',
@@ -8,9 +7,6 @@ document.addEventListener(
     const element = document.getElementById('app');
     renderComponent(MyComponent, {
       element: element!,
-      meta: {
-        locale: new LocaleService('en_US'),
-      },
     });
   },
   { once: true }
